requirements.txt
fastapi
sqlmodel
uvicorn[basic]
python-dotenv
bcrypt
psycopg2-binary
email-validator
python-jose[cryptography]